src/requirements.txt
Django==3.2.0
djangorestframework==3.12.4
psycopg2==2.8.4
python-decouple==3.6.0
